Removing Google Analytics. Reason in description
REASON: I added Google Analytics hoping to replace it to a FOSS alternative. However, all of the options are thought to be held in a server, no in GitHub Pages static site. In addition, I decided that I don't need to check if I have an audience of 5 or 50 readers: I write just for the pleasing act of writing, not to attract audience. That being said, I reserve myself the right to add it in the future (ofc with a proper reason)